+++ examples/logic_synthesis/blif/test_cgp_ls_blif.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
-from src.gp.tiny_cgp import *
-from src.gp.problem import Problem
-from src.gp.tinyverse import Var
+from gp.tiny_cgp import *
+from gp.problem import Problem
+from gp.tinyverse import Var
 import requests
 
 # from dd.cudd import BDD
@@ -10,8 +10,8 @@ from dd.autoref import BDD, Function
 
 Function.__xor__ = lambda self, other: self._apply('xor', other)
 
-from src.benchmark.logic_synthesis.blif_parser.blif import BlifFile
-from src.gp.tinyverse import Function
+from benchmark.logic_synthesis.blif_parser.blif import BlifFile
+from gp.tinyverse import Function
 
 AND = Function(2, 'AND', lambda x, y: x & y)
 OR = Function(2, 'OR', lambda x, y: x | y)
