+++ examples/policy_learning/test_cgp_pl_ale.py
@@ -14,11 +14,11 @@ Action space: Discrete(6)
 
 Observation space: Box(0, 255, (210, 160, 3), uint8)
 """
-from src.benchmark.policy_search.pl_benchmark import PLBenchmark, ALEArgs
-from src.gp.tiny_cgp import *
+from benchmark.policy_search.pl_benchmark import PLBenchmark, ALEArgs
+from gp.tiny_cgp import *
 import gymnasium as gym
-from src.gp.problem import PolicySearch
-from src.gp.functions import *
+from gp.problem import PolicySearch
+from gp.functions import *
 import warnings
 import numpy
