+++ examples/program_synthesis/test_cgp_ps.py
@@ -8,12 +8,12 @@ https://leetcode.com/problems/power-of-two/description/
 
 """
 
-from src.gp.tiny_cgp import *
-from src.gp.problem import ProgramSynthesis
-from src.benchmark.program_synthesis.ps_benchmark import PSBenchmark
-from src.benchmark.program_synthesis.leetcode.power_of_two import *
-from src.gp.functions import *
-from src.gp.tinyverse import Var, Const
+from gp.tiny_cgp import *
+from gp.problem import ProgramSynthesis
+from benchmark.program_synthesis.ps_benchmark import PSBenchmark
+from benchmark.program_synthesis.leetcode.power_of_two import *
+from gp.functions import *
+from gp.tinyverse import Var, Const
 
 NUM_INPUTS = 1
 functions = [ADD, SUB, MUL, DIV, AND, OR, NAND, NOR, NOT, IF, LT, GT]
