+++ examples/symbolic_regression/test_cgp_sr.py
@@ -12,12 +12,12 @@ no serious benchmark. It only serves as an example for SR as an application
 domain for TinyverseGP:
 """
 
-from src.gp.tiny_cgp import *
-from src.gp.problem import BlackBox
-from src.benchmark.symbolic_regression.sr_benchmark import SRBenchmark
-from src.gp.functions import *
-from src.gp.loss import *
-from src.gp.tinyverse import Var, Const
+from gp.tiny_cgp import *
+from gp.problem import BlackBox
+from benchmark.symbolic_regression.sr_benchmark import SRBenchmark
+from gp.functions import *
+from gp.loss import *
+from gp.tinyverse import Var, Const
 
 functions = [ADD, SUB, MUL, DIV]
 terminals = [Var(0), Const(1)]
