+++ examples/symbolic_regression/test_tgp_sr.py
@@ -12,11 +12,11 @@ no serious benchmark. It only serves as an example for SR as an application
 domain for TinyverseGP:
 """
 
-from src.gp.tiny_tgp import *
-from src.gp.functions import *
-from src.gp.loss import *
-from src.gp.problem import BlackBox
-from src.benchmark.symbolic_regression.sr_benchmark import SRBenchmark
+from gp.tiny_tgp import *
+from gp.functions import *
+from gp.loss import *
+from gp.problem import BlackBox
+from benchmark.symbolic_regression.sr_benchmark import SRBenchmark
 
 config = GPConfig(
     num_jobs=1,
