+++ src/benchmark/logic_synthesis/ls_benchmark.py
@@ -2,7 +2,7 @@
 Benchmark representation module for logic synthesis.
 """
 
-from src.benchmark.benchmark import Benchmark
+from benchmark.benchmark import Benchmark
 import src.benchmark.logic_synthesis.boolean_benchmark_tools.benchmark_reader as BenchmarkReader
 
 class LSBenchmark(Benchmark):
