+++ src/gp/tiny_lgp.py
@@ -15,8 +15,8 @@ import time
 import operator
 from dataclasses import dataclass
 from collections import namedtuple
-from src.gp.functions import Function
-from src.gp.tinyverse import GPModel, Hyperparameters, GPConfig
+from gp.functions import Function
+from gp.tinyverse import GPModel, Hyperparameters, GPConfig
 
 try:
     from icecream import ic
