+++ src/gp/tiny_tgp.py
@@ -6,8 +6,8 @@ TinyGP: A minimalistic implementation of tree-based Genetic Programming for
 import random
 import copy
 import time
-from src.gp.problem import *
-from src.gp.tinyverse import *
+from gp.problem import *
+from gp.tinyverse import *
 
 class Node:
     '''
